+++ src/hivememory/agent_runtime/aliases/__init__.py
@@ -1,0 +1,6 @@
+"""Agent Runtime 的 alias 缓存与解析能力。"""
+
+from hivememory.agent_runtime.aliases.cache import KoakumaAtomCache
+from hivememory.agent_runtime.aliases.resolver import ResolveResult, RuntimeAliasResolver
+
+__all__ = ["KoakumaAtomCache", "ResolveResult", "RuntimeAliasResolver"]
