+++ src/hivememory/agent_runtime/execution/__init__.py
@@ -1,0 +1,6 @@
+"""单 frame 执行循环与模型生成适配器。"""
+
+from hivememory.agent_runtime.execution.loop import AgentLoopExecutor
+from hivememory.agent_runtime.execution.worker import WorkerAgentService
+
+__all__ = ["AgentLoopExecutor", "WorkerAgentService"]
