+++ src/hivememory/agent_runtime/mtp/__init__.py
@@ -2,7 +2,7 @@
 MTP runtime 子包。
 """
 
-from hivememory.agent_runtime.mtp.mtp_executor import KoakumaMTPExecutor, MTPExecutor
+from hivememory.agent_runtime.mtp.executor import KoakumaMTPExecutor, MTPExecutor
 from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
 
 __all__ = [
