+++ tests/e2e/pipeline/test_kernel_loop_e2e.py
@@ -27,7 +27,7 @@ import pytest
 from unittest.mock import MagicMock
 
 from hivememory.system.config import LLMConfig, KoakumaConfig
-from hivememory.agent_runtime.worker_agent import WorkerAgentService
+from hivememory.agent_runtime.execution import WorkerAgentService
 from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
 from hivememory.core.protocol.models import AgentRunResult, RetrievalResponse
 from hivememory.prompts.mtp import MTPPromptBuilder
