+++ tests/unit/agent_runtime/aliases/test_resolver.py
@@ -2,10 +2,10 @@ import pytest
 from unittest.mock import MagicMock
 from uuid import uuid4
 
-from hivememory.agent_runtime.cache import KoakumaAtomCache
+from hivememory.agent_runtime.aliases.cache import KoakumaAtomCache
 from hivememory.agent_runtime.models import MTPExecutionContext
 from hivememory.agent_runtime.pending_atom import PendingAtomRuntime
-from hivememory.agent_runtime.resolver import RuntimeAliasResolver
+from hivememory.agent_runtime.aliases.resolver import RuntimeAliasResolver
 from hivememory.core.models import (
     Identity,
     IndexLayer,
