+++ tests/unit/agent_runtime/execution/__init__.py
@@ -1,0 +1,1 @@
+"""Frame execution tests."""
